+++ docs/changelog.md
@@ -11,6 +11,22 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Replaced OrderedDict with dict.
 - Deprecated support for Python versions <= 3.6.
 
+## [0.35.7] -- 2023-07-19
+### Fixed 
+- incorrect setting of sample and subsample indexes using from_dict function (#452)
+- clarified debug messages
+
+## [0.35.6] -- 2023-06-27
+### Added
+- `orient` argument to `to_dict` method
+
+### Fixed
+- The name of the raw subsample object to match the actual name (list). Commit: #442
+
+### Changed
+- Reduced the number of items returned in the to_dict(extended=True) method to 3, with the name and description now stored in the config key.
+
+
 ## [0.35.5] -- 2023-03-27
 ### Fixed
 - A [bug](https://github.com/pepkit/peppy/issues/435) with custom sample ids